requirements.txt
flask
openai
tavily-python
python-dotenv
gunicorn